Adding detailed comments to test
This logic took me ages to work out so I've added some extremely detailed documentation to the fixture. This is not the right thing to do or the right place to do it - the code and tests should be refactored so this logic is clear from reading the tests; however, I do not want to make too many changes at once so this will do for now.