Removing errant +1
I can't see why this +1 has been added. The work it is doing means we include
the current time period. Pending more complex logic for handling this (to allow
users to enter data for previous time periods but not current time period) I do
not think we should collect data for the current time period.